+++ pkg/utils/instance.go
@@ -1,0 +1,7 @@
+package utils
+
+import "fmt"
+
+func GenerateInstanceName(clusterID string, nodeClaimName string) string {
+	return fmt.Sprintf("%s-%s", clusterID, nodeClaimName)
+}