+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Unreleased
 
 - Add Karpenter deployment manifests
 - provider: drop clusterName (unused attribute)
+- cloudprovider: use clusterID instead of clusterName
 
 0.0.1
 ------